make longest and shortest path finders weight aware

+++ src/Chayka/GraphBuilder/DefaultGraphBuilderExtensions.cs
@@ -7,7 +7,7 @@
             return builder.AddVertex(new DefaultVertex<T>(vertex));
         }
 
-        public static IGraphBuilder<T> AddEdge<T>(this IGraphBuilder<T> builder, T source, T target, int weight = 1)
+        public static IGraphBuilder<T> AddEdge<T>(this IGraphBuilder<T> builder, T source, T target, double weight = 1)
         {
             return builder.AddEdge(new DefaultEdge<IVertex<T>>(new DefaultVertex<T>(source), new DefaultVertex<T>(target), weight));
         }
